refactoring folder hierarchy
