--- imgui.ini
@@ -1,45 +1,0 @@
-[Window][Debug##Default]
-Pos=60,60
-Size=400,400
-Collapsed=0
-
-[Window][Time Elapsed]
-Pos=5,5
-Size=345,110
-Collapsed=0
-
-[Window][Distance]
-Pos=5,120
-Size=345,110
-Collapsed=0
-
-[Window][Calories Burnt]
-Pos=355,5
-Size=345,110
-Collapsed=0
-
-[Window][Speed]
-Pos=355,120
-Size=345,110
-Collapsed=0
-
-[Window][Heart Rate]
-Pos=705,5
-Size=305,225
-Collapsed=0
-
-[Window][Separator]
-Pos=5,235
-Size=1005,36
-Collapsed=0
-
-[Window][ImGui Demo]
-Pos=297,-27
-Size=550,680
-Collapsed=0
-
-[Window][Shite]
-Pos=5,5
-Size=768,480
-Collapsed=0
-
