--- notes
@@ -1,13 +1,0 @@
-With Graph:
-    Power,
-    Heart Rate
-
-Without Graph:
-    Time,
-    
-    Speed,
-    Distance,
-    Calories
-
-    Voltage,
-    Current,
